+++ packages/web-platform/src/libs/extension/types.ts
@@ -1,0 +1,11 @@
+export type ExtensionOperationTypes = "clientHello";
+
+export interface ClientHello {
+    windowId: string;
+}
+
+export interface ClientHelloResponse {
+    widget: {
+        inject: boolean;
+    };
+}