+++ packages/web/src/appManager/controller.ts
@@ -82,6 +82,7 @@ export class AppManagerController implements LibController {
 
     public async startApplication(appName: string, context?: object, options?: Glue42Web.AppManager.ApplicationStartOptions): Promise<Glue42Web.AppManager.Instance> {
         // reuseId is a hidden property for workspaces-only use
+        // forceChromeTab is a hidden property for extension use only
         const startOptions: ApplicationStartConfig = {
             name: appName,
             waitForAGMReady: options?.waitForAGMReady ?? true,
@@ -92,7 +93,8 @@ export class AppManagerController implements LibController {
             height: options?.height,
             relativeTo: options?.relativeTo,
             relativeDirection: options?.relativeDirection,
-            id: (options as any)?.reuseId
+            id: (options as any)?.reuseId,
+            forceChromeTab: (options as any)?.forceTab
         };
 
         const openResult = await this.bridge.send<ApplicationStartConfig, InstanceData>("appManager", operations.applicationStart, startOptions);