+++ packages/fdc3/changelog.md
@@ -1,3 +1,7 @@
+2.4.0
+feat: added support for GDX (Glue42 Developer Extension)
+2.3.5
+chore: bump due to dependencies update
 2.3.4
 chore: bump due to dependencies update
 2.3.3