+++ packages/workspaces-api/changelog.md
@@ -1,3 +1,5 @@
+1.10.0
+feat: added support for GDX (Glue42 Developer Extension)
 1.9.1
 chore: bump due to dependencies update
 1.9.0